requirements.txt
Flask>=3.0
python-dotenv>=1.0
openai>=1.40
PyPDF2>=3.0
